fix import and rename chart
