+++ ts/webui/README.md
@@ -7,6 +7,8 @@ WebUI is built by using [React](https://reactjs.org/docs/getting-started.html) a
 
 * Please refer the [installation doc](https://github.com/microsoft/nni#installation) to run an experiment.
 
+* Please set `baseUrl` value `.` in `tsconfig.json` file.
+
 * Use this command in `webui/ts` directory when you change webui code. And then refresh website to see latest pages.
     ```bash
     yarn build
@@ -15,5 +17,7 @@ WebUI is built by using [React](https://reactjs.org/docs/getting-started.html) a
 ## PR
 
 * WebUI uses [eslint](https://eslint.org/docs/user-guide/getting-started) and [prettier](https://prettier.io/docs/en/index.html) to format code. You could use the command `yarn sanity-check` to check the code error status. And use `yarn sanity-check --fix` could modifiy the most code style error before you send PR.
+
+* Change `baseUrl` value back.
 
 * You could send the PR if `yarn release` gets successful build after formatting code.
