+++ ts/webui/src/components/experiment/trialdetail/chart/DefaultMetricPoint.tsx
