+++ ts/webui/src/components/experiment/trialdetail/chart/Duration.tsx
